requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
neutron-lib>=1.18.0 # Apache-2.0
oslo.config>=5.2.0 # Apache-2.0
oslo.i18n>=3.15.3 # Apache-2.0
oslo.log>=3.36.0 # Apache-2.0
oslo.utils>=3.33.0 # Apache-2.0
oslo.messaging>=5.29.0 # Apache-2.0
pbr!=2.1.0,>=2.0.0 # Apache-2.0
python-ironicclient>=2.3.0 # Apache-2.0
tooz>=1.58.0 # Apache-2.0
neutron>=13.0.0.0b1  # Apache-2.0
keystoneauth1>=3.4.0

requests>=2.18.4
six>=1.11.0
